[axe.c] Use better source of fwd/rev barcode seqs
We now go thru every barcode pair, and if the fwd seq is not in the 0
mismatch trie, we add it to both the 0mm trie, mutate it and then add
the mutated seqs to their tries. If it's already in the trie, we skip
this pair.

Then, we do exactly the same for the reverse barcode seqs.

This guarntees that ALL forward and reverse barcodes are read.